Add files via upload
I made this project for my final assignment in electronics. the parts you need are: Arduino Uno, Breadboard, HC-SR04 ultrasonic sensor with 4 pins, a servomotor, a red and green led. an LCD sreen and 2 potentiometers.

to start, connect the ultrasonic sensor; the left pin goes to the 5volt, 2nd pin (trigpin) goes to pin 12 on the arduino, then the 3rd pin of the sensor (echopin) goes to pin 13 of the arduino, connect the last pin to the ground. Now connect the LEDs. the short legs both go to the ground. the other leg of the red led goes to 4, the green led goes to pin 3. You connect both to the arduino. We will use the first potentiometer to set the range. the pin with the ohm sign on is connected to the ground, the middle pin is connected to the A0 and the last pin to the 5V. We will use the 2nd potentiometer to adjust the LCD screen. you connect the 2 outermost pins in the same way as the previous one. the middle pin is connected with the 3rd pin of the LCD.then we will now connect the other pins of the LCD screen. you connect the first pin to the ground, the second to the 5V. You normally have the 3rd connected to the potentiometer. the 4th pin you connect to pin 9 of the arduino, the 5th pin you connect to the ground. we are going to connect pin 6 to pin 8 of the arduino. now we are going to skip 4 pins to connect pin 11 to pin 7 of the arduino. pin 12 goes to pin 6.The 13th goes to pin 5 of the arduino to connect pin 14 to pin 4. you connect the penultimate pin to the 5V and the last to the ground. when you download the program normaly you will see how far away something is. you can read that on the LCD screen.